Extract storybook mock handlers into separate files
Split the inline handler definitions in the .storybook handlers.ts
files of @comet/admin and the root storybook into one file per
handler, so handlers.ts only wires them together.
